add spectral test and also implement supersampling
