+++ scene/spectrum-test.rt
@@ -1,0 +1,10 @@
+A 0.1 255,255,255
+
+C 0,-0.1,-4 0,0,1 25
+
+L  -0.600,-0.400,-2.100   0.5      255,010,010
+L  +0.000,+0.600,-2.100   0.5      010,255,010
+L  +0.600,-0.400,-2.100   0.5      010,010,255
+
+pl 0,0,25 0,0,-1 255,255,255 0
+sp 0,0,10 1 255,255,255 100
