+++ scene/test.rt
@@ -1,6 +1,6 @@
 A 0.2 255,255,255
 
-C 0,0,0 0,0,1 120
+C 0,0,0 0,0,1 80
 
 L 0,0,10 0.4 255,255,255
 L -10,0,0 0.2 255,0,0
